Aula 91

+++ src/main/java/io/github/kauabetineli/libraryapi/service/AutorService.java
@@ -1,9 +1,12 @@
 package io.github.kauabetineli.libraryapi.service;
 
 import io.github.kauabetineli.libraryapi.controller.dto.AutorDTO;
+import io.github.kauabetineli.libraryapi.exceptions.OperacaoNaoPermitidaException;
 import io.github.kauabetineli.libraryapi.model.Autor;
 import io.github.kauabetineli.libraryapi.repository.AutorRepository;
+import io.github.kauabetineli.libraryapi.repository.LivroRepository;
 import io.github.kauabetineli.libraryapi.validator.AutorValidator;
+import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -11,15 +14,12 @@ import java.util.Optional;
 import java.util.UUID;
 
 @Service
+@RequiredArgsConstructor
 public class AutorService {
 
     private final AutorRepository autorRepository;
     private final AutorValidator autorValidator;
-
-    public AutorService(AutorRepository autorRepository, AutorValidator autorValidator) {
-        this.autorRepository = autorRepository;
-        this.autorValidator = autorValidator;
-    }
+    private final LivroRepository livroRepository;
 
     public Autor salvar(Autor autor) {
         autorValidator.validar(autor);
@@ -39,6 +39,10 @@ public class AutorService {
     }
 
     public void deletar(Autor autor) {
+        if(possuiLivro(autor)){
+            throw new OperacaoNaoPermitidaException(
+                    "Não é permitido excluir um Autor que possui livros cadastrados!");
+        }
         autorRepository.delete(autor);
     }
 
@@ -57,6 +61,10 @@ public class AutorService {
 
         return autorRepository.findAll();
 
+    }
+
+    public boolean possuiLivro(Autor autor){
+        return livroRepository.existsByAutor(autor);
     }
 
 }
